+++ example/a.ts
@@ -4,12 +4,24 @@ const app = new Elysia()
 	.setModel({
 		number: t.Number()
 	})
-	.post('/', ({ body: { data } }) => data, {
-		schema: {
-			response: 'number',
-			body: t.Object({
-				data: t.Number()
-			})
+	.post(
+		'/',
+		({ set, body: { status, response } }) => {
+			set.status = status
+
+			return response
+		},
+		{
+			schema: {
+				body: t.Object({
+					status: t.Number(),
+					response: t.Any()
+				}),
+				response: {
+					200: t.String(),
+					201: t.Number()
+				}
+			}
 		}
-	})
+	)
 	.listen(8080)